+++ src/refactorings/extract-function/extract-function.test.ts
@@ -114,6 +114,39 @@ describe("Extract Function", () => {
     });
   });
 
+  it("should expand selection to closest list of statements", () => {
+    const { code, selection } =
+      new InMemoryEditor(`async function bookRoom(customerId, date) {
+  const custo[start]mer = await findCustomer(customerId);
+  if (!customer) {
+    throw new Error("Customer not found");[end]
+  }
+
+  logger.debug("Room was booked");
+}`);
+
+    const result = extractFunction({
+      state: "new",
+      code,
+      selection
+    });
+
+    const { selection: expectedSelection } =
+      new InMemoryEditor(`async function bookRoom(customerId, date) {
+  [start]const customer = await findCustomer(customerId);
+  if (!customer) {
+    throw new Error("Customer not found");
+  }[end]
+
+  logger.debug("Room was booked");
+}`);
+    expect(result).toEqual({
+      action: "delegate",
+      command: "extract function",
+      selection: expectedSelection
+    });
+  });
+
   it("should show an error if the editor does not support the refactoring", () => {
     const { code, selection } = new InMemoryEditor();